+++ Adapters/DealAdapter.php
@@ -6,21 +6,29 @@ class DealAdapter
     public function addDeal($date, $store)
     {
         $query = "INSERT INTO `deal` (`date`,`store`) VALUES ('$date','$store')";
-        return DataBase::queryINSERT($query);
+        DataBase::queryINSERT($query);
     }
 
     public function deleteDeal($id)
     {
-        $query = "DELETE FROM `deal` WHERE id =". $id;
-        DataBase::query($query);
+        $resault = false;
+        if(!empty($id)) {
+            $query = "DELETE FROM `deal` WHERE id =". $id;
+            DataBase::query($query);
+        }
+        return $resault;
     }
 
     public function editDeal(array $arr,$id)
     {
-        $set = implode(',', $arr);
-        $query = "UPDATE `deal` 
+        $resault = false;
+        if(!empty($id)) {
+            $set = implode(',', $arr);
+            $query = "UPDATE `deal` 
                     SET ".$set." WHERE id = ".$id;
-        DataBase::query($query);
+            DataBase::query($query);
+        }
+        return $resault;
     }
 
     public function DealDate($id)
@@ -35,5 +43,15 @@ class DealAdapter
         $query = "SELECT store FROM deal WHERE id =". $id;
         $store = DataBase::query($query);
         return $store[0]['store'];
+    }
+
+    public function deleteCustomerDealBook($id)
+    {
+        $resault = false;
+        if(!empty($id)) {
+            $query = "DELETE FROM `CustomerDealBook` WHERE id_deal =". $id;
+            database::query($query);
+        }
+        return $resault;
     }
 }
